requirements/requirements.txt
asgiref==3.8.1
babel==2.16.0
Django==5.1.3
django-allauth==65.2.0
django-cors-headers==4.6.0
django-countries==7.6.1
django-money==3.5.3
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
psycopg2==2.9.10
py-moneyed==3.0
pycountry==24.6.1
PyJWT==2.9.0
python-decouple==3.8
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
